pyproject.toml
[tool.poetry]
name = "bcg-spark-assessment"
version = "0.1.0"
description = ""
authors = ["Prajwal R"]
packages = [
    { include = "src"}
]

[tool.poetry.dependencies]
python = "^3.9"
pyspark = "3.5.0"
loguru = "^0.7.0"
pyyaml = "^6.0.0"
pandas = "^2.2.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
